Undo size_t patch

+++ include/caffe/data_layers.hpp
@@ -280,7 +280,7 @@ class MemoryDataLayer : public BaseDataLayer<Dtype> {
 
   // Reset should accept const pointers, but can't, because the memory
   //  will be given to Blob, which is mutable
-  void Reset(Dtype* data, Dtype* label, size_t n);
+  void Reset(Dtype* data, Dtype* label, int n);
   void set_batch_size(int new_size);
 
   int batch_size() { return batch_size_; }
@@ -295,7 +295,7 @@ class MemoryDataLayer : public BaseDataLayer<Dtype> {
   int batch_size_, channels_, height_, width_, size_;
   Dtype* data_;
   Dtype* labels_;
-  size_t n_;
+  int n_;
   size_t pos_;
   Blob<Dtype> added_data_;
   Blob<Dtype> added_label_;
